feat: 🎸 commom add help
