+++ packages/core/docs/component.md
@@ -105,84 +105,6 @@ class ComponentDemo {
 }
 ```
 
-#### help
-
-A simple, data-driven module for creating a usage guide. 具体使用请查看[文档](https://github.com/75lb/command-line-usage)
-
-```typescript
-const { Component, IComponent } = require('@serverless-devs/core');
-
-class ComponentDemo {
-  @HComponent() component: IComponent;
-
-  help() {
-    const sections = [
-      {
-        header: 'A typical app',
-        content: 'Generates something {italic very} important.',
-      },
-      {
-        header: 'Options',
-        optionList: [
-          {
-            name: 'input',
-            typeLabel: '{underline file}',
-            description: 'The input to process.',
-          },
-          {
-            name: 'help',
-            description: 'Print this usage guide.',
-          },
-        ],
-      },
-      {
-        header: 'Examples',
-        content: [
-          {
-            desc: '1. A concise example. ',
-            example: '$ example -t 100 lib/*.js',
-          },
-          {
-            desc: '2. A long example. ',
-            example: '$ example --timeout 100 --src lib/*.js',
-          },
-          {
-            desc:
-              '3. This example will scan space for unknown things. Take cure when scanning space, it could take some time. ',
-            example:
-              '$ example --src galaxy1.facts galaxy1.facts galaxy2.facts galaxy3.facts galaxy4.facts galaxy5.facts',
-          },
-        ],
-      },
-    ];
-    this.component.help(sections);
-  }
-}
-```
-
-![Demo](https://img.alicdn.com/imgextra/i2/O1CN01QGUzRz1qIDvq8gUPR_!!6000000005472-1-tps-1337-221.gif)
-
-输出数据
-
-```
-
-A typical app
-
-  Generates something very important.
-
-Options
-
-  --input file    The input to process.
-  --help string   Print this usage guide.
-
-Examples
-
-  1. A concise example.                                                                                          $ example -t 100 lib/*.js
-  2. A long example.                                                                                             $ example --timeout 100 --src lib/*.js
-  3. This example will scan space for unknown things. Take cure when scanning space, it could take some time.    $ example --src galaxy1.facts galaxy1.facts galaxy2.facts galaxy3.facts galaxy4.facts galaxy5.facts
-
-```
-
 ## credentials
 
 用于获取密钥信息, 目前 Provider 支持 [alibaba/aws/azure/baidu/google/huawei/tencent/custom]
