+++ packages/core/src/common/index.ts
@@ -4,3 +4,4 @@ export { default as spinner } from './spinner';
 export { default as zip } from './zip';
 export { default as unzip } from 'decompress';
 export { default as i18n } from './i18n';
+export { default as help } from './help';
